backend/services/process_emails/.ipynb_checkpoints/requirements-checkpoint.txt
Flask
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-firestore
google-cloud-pubsub
gunicorn